Add additional UID check
Example:

```
$ id -u
1000

$ runas -s -u 1000
```

The above will a new shell as the current user, thereby bypassing authentication. By default 'run_argv_args[]' will contain 'uid=0' which is updated by the option '--uid'. But if there is anything wrong with '(int) run_argc_args' this may not be updated in the systemd command arguments. However it will be the UID that being authenticated. This means that systemd will be invoked using the following:

```
systemd-run --uid 0 ...
```

So let's make sure that the UID being authenticated is the same as the one being passed to systemd by making sure that nothing is wrong with 'run_argc_args' or 'run_argv_args[]' for that mater. It's a simple check so no reason not to.